+++ src/main/java/calculator/delimiter/domain/Delimiter.java
@@ -1,4 +1,4 @@
-package calculator.delimiter;
+package calculator.delimiter.domain;
 
 public record Delimiter(String value) {
